refactor: /components dir structure
