+++ src/components/rjsf/CustomUISchema.js
@@ -1,4 +1,4 @@
-export const uiSchema = {
+export const CustomUISchema = {
   'ui:options': {
     timeFormat: 'hh:mm:ss',
     'ui:placeholder': '00:00:00'
