+++ src/components/rjsf/CustomWidgets.js
@@ -84,4 +84,4 @@ CustomTextWidget.propTypes = {
   schema: PropTypes.object
 }
 
-export const widgets = { checkbox: CustomCheckboxWidget, time: CustomTimeWidget, text: CustomTextWidget }
+export const CustomWidgets = { checkbox: CustomCheckboxWidget, time: CustomTimeWidget, text: CustomTextWidget }
